Clean up the usage of the event hook for damage pipeline
