--- module/pipelines/apply-damage.mjs
@@ -1,163 +1,0 @@
-import { SYSTEM } from '../helpers/config.mjs';
-import { Flags } from '../helpers/flags.mjs';
-import { ChecksV2 } from '../checks/checks-v2.mjs';
-import { CheckConfiguration } from '../checks/check-configuration.mjs';
-import { DamageCustomizer } from './damage-customizer.mjs';
-import { getSelected, getTargeted } from '../helpers/target-handler.mjs';
-import { DamagePipeline, DamageRequest } from './damage-pipeline.mjs';
-import { InlineSourceInfo } from '../helpers/inline-helper.mjs';
-
-// Typedefs
-
-/**
- * @typedef ClickModifiers
- * @prop {boolean} alt
- * @prop {boolean} ctrl
- * @prop {boolean} shift
- */
-
-/**
- * @callback DamageModifier
- * @param {number} baseDamage
- * @param {ClickModifiers} modifiers
- * @return {number}
- */
-
-/**
- * @typedef BeforeApplyHookData
- * @prop {Event | null} event
- * @prop {FUActor[]} targets
- * @prop {string | null} sourceUuid
- * @prop {string | null} sourceName
- * @prop {import('../helpers/typedefs.mjs').BaseDamageInfo} baseDamageInfo
- * @prop {import('./damage-customizer.mjs').ExtraDamageInfo} extraDamageInfo
- * @prop {ClickModifiers | null} clickModifiers
- */
-
-/**
- * @typedef ApplyTargetHookData
- * @prop {FUActor} target
- * @prop {string | null} sourceUuid
- * @prop {string | null} sourceName
- * @prop {DamageType} damageType
- * @prop {number} total
- * @prop {ClickModifiers | null} clickModifiers
- * @prop {ExtraDamageInfo} extraDamageInfo
- * @prop {ApplyTargetOverrides} overrides
- */
-
-/**
- * @typedef ApplyTargetOverrides
- * @prop {number | null} affinity
- * @prop {number | null} total
- */
-
-function attachDamageApplicationHandler(message, jQuery) {
-	const check = message.getFlag(SYSTEM, Flags.ChatMessage.CheckParams);
-	let sourceUuid = null;
-	let sourceName;
-	let baseDamageInfo;
-	let disabled = false;
-
-	if (check && check.damage) {
-		sourceName = check.details.name;
-		baseDamageInfo = {
-			total: check.damage.total,
-			type: check.damage.type,
-			modifierTotal: check.damage.modifierTotal,
-		};
-	}
-
-	if (ChecksV2.isCheck(message)) {
-		const damage = CheckConfiguration.inspect(message).getDamage();
-		if (damage) {
-			sourceUuid = message.getFlag(SYSTEM, Flags.ChatMessage.CheckV2)?.itemUuid;
-			sourceName = message.getFlag(SYSTEM, Flags.ChatMessage.Item)?.name;
-			baseDamageInfo = {
-				total: damage.total,
-				type: damage.type,
-				modifierTotal: damage.modifierTotal,
-			};
-		}
-	}
-
-	const handleClick = async (event, getTargetsFunction) => {
-		event.preventDefault();
-		if (!disabled) {
-			disabled = true;
-			const targets = await getTargetsFunction(event);
-			if (event.ctrlKey || event.metaKey) {
-				DamageCustomizer(
-					baseDamageInfo,
-					targets,
-					(extraDamageInfo) => {
-						handleDamageApplication(event, targets, sourceUuid, sourceName, baseDamageInfo, extraDamageInfo);
-						disabled = false;
-					},
-					() => {
-						disabled = false;
-					},
-				);
-			} else {
-				handleDamageApplication(event, targets, sourceUuid, sourceName, baseDamageInfo, {});
-				disabled = false;
-			}
-		}
-	};
-
-	jQuery.find(`a[data-action=applySingleDamage]`).click((event) => handleClick(event, getSingleTarget));
-	jQuery.find(`a[data-action=applySelectedDamage]`).click((event) => handleClick(event, getSelected));
-	jQuery.find(`a[data-action=applyTargetedDamage]`).click((event) => handleClick(event, getTargeted));
-	jQuery.find(`a[data-action=selectDamageCustomizer]`).click(async (event) => {
-		if (!disabled) {
-			disabled = true;
-			const targets = await getTargeted(event);
-			DamageCustomizer(
-				baseDamageInfo,
-				targets,
-				(extraDamageInfo) => {
-					handleDamageApplication(event, targets, sourceUuid, sourceName, baseDamageInfo, extraDamageInfo);
-					disabled = false;
-				},
-				() => {
-					disabled = false;
-				},
-			);
-		}
-	});
-}
-
-/**
- *
- * @param {Event} event
- * @param {FUActor[]} targets
- * @param {string} sourceUuid
- * @param {string} sourceName
- * @param {import('../helpers/typedefs.mjs').BaseDamageInfo} baseDamageInfo
- * @param {import('./damage-customizer.mjs').ExtraDamageInfo} extraDamageInfo
- * @returns {void}
- */
-async function handleDamageApplication(event, targets, sourceUuid, sourceName, baseDamageInfo, extraDamageInfo) {
-	const sourceInfo = new InlineSourceInfo(sourceName, sourceUuid, null);
-	const request = new DamageRequest(sourceInfo, targets, baseDamageInfo, extraDamageInfo);
-	request.setEvent(event);
-	await DamagePipeline.process(request);
-}
-
-/**
- * @param {Event} event
- * @returns {FUActor[]}
- */
-function getSingleTarget(event) {
-	const dataId = $(event.target).closest('a').data('id');
-	const actor = fromUuidSync(dataId);
-	if (!actor) {
-		ui.notifications.warn('FU.ChatApplyEffectNoActorsTargeted', { localize: true });
-		return [];
-	}
-	return [actor];
-}
-
-export function registerChatInteraction() {
-	Hooks.on('renderChatMessage', attachDamageApplicationHandler);
-}
